Rename script to numberChecker
