+++ 03_numberChecker/numberChecker.js
