+++ 03_numberChecker/solution/numberChecker-solution.js
